requirements.txt
#pip install langchain langchain-chroma chromadb sentence-transformers
##pip install langchain-community
#pip install -U pymilvus>=2.4.2
#pip install mem0ai

langchain>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-milvus>=0.1.0
langchain-huggingface>=0.0.1
langchain-groq>=0.1.0
chromadb>=0.4.0
pymilvus>=2.4.2
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
rich>=13.0.0
numpy>=1.24.0
mem0ai

